+++ pkg/cmd/delete/clusterpoolhost/exec.go
@@ -1,5 +1,5 @@
 // Copyright Contributors to the Open Cluster Management project
-package use
+package clusterpoolhost
 
 import (
 	"fmt"
@@ -10,7 +10,7 @@ import (
 
 func (o *Options) complete(cmd *cobra.Command, args []string) (err error) {
 	if len(args) < 1 {
-		return fmt.Errorf("clustername is missing")
+		return fmt.Errorf("clusterpoolcph name is missing")
 	}
 	o.ClusterHostPool = args[0]
 	return nil
@@ -21,5 +21,17 @@ func (o *Options) validate() error {
 }
 
 func (o *Options) run() (err error) {
-	return clusterpoolhost.VerifyContext(o.ClusterHostPool, o.CMFlags.DryRun, o.outputFile)
+	cphs, err := clusterpoolhost.GetClusterPoolHosts()
+	if err != nil {
+		return err
+	}
+	cph, err := cphs.GetClusterPoolHost(o.ClusterHostPool)
+	if err != nil {
+		return err
+	}
+	err = cph.DeleteClusterPoolHost()
+	if err != nil {
+		return err
+	}
+	return nil
 }