+++ pkg/cmd/delete/clusterpoolhost/options.go
@@ -1,5 +1,5 @@
 // Copyright Contributors to the Open Cluster Management project
-package use
+package clusterpoolhost
 
 import (
 	genericclioptionscm "github.com/open-cluster-management/cm-cli/pkg/genericclioptions"
@@ -10,8 +10,6 @@ type Options struct {
 	//CMFlags: The generic optiosn from the cm cli-runtime.
 	CMFlags         *genericclioptionscm.CMFlags
 	ClusterHostPool string
-	//The file to output the resources will be sent to the file.
-	outputFile string
 }
 
 func newOptions(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOStreams) *Options {