+++ pkg/cmd/delete/cmd.go
@@ -3,6 +3,7 @@ package delete
 
 import (
 	"github.com/open-cluster-management/cm-cli/pkg/cmd/delete/cluster"
+	"github.com/open-cluster-management/cm-cli/pkg/cmd/delete/clusterpoolhost"
 	genericclioptionscm "github.com/open-cluster-management/cm-cli/pkg/genericclioptions"
 	clusteradmdeletetoken "open-cluster-management.io/clusteradm/pkg/cmd/delete/token"
 	genericclioptionsclusteradm "open-cluster-management.io/clusteradm/pkg/genericclioptions"
@@ -20,5 +21,6 @@ func NewCmd(clusteradmFlags *genericclioptionsclusteradm.ClusteradmFlags, cmFlag
 
 	cmd.AddCommand(cluster.NewCmd(cmFlags, streams))
 	cmd.AddCommand(clusteradmdeletetoken.NewCmd(clusteradmFlags, streams))
+	cmd.AddCommand(clusterpoolhost.NewCmd(cmFlags, streams))
 	return cmd
 }