+++ pkg/cmd/setcph/cmd.go
@@ -1,5 +1,5 @@
 // Copyright Contributors to the Open Cluster Management project
-package use
+package setcph
 
 import (
 	"fmt"
@@ -14,15 +14,15 @@ import (
 
 var example = `
 # Use a cluster
-%[1]s use cluster
+%[1]s set-cph clusterpoolhost
 `
 
-// NewCmd provides a cobra command for using a cluster claim
+// NewCmd provides a cobra command to use a clusterpoolhost
 func NewCmd(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOStreams) *cobra.Command {
 	o := newOptions(cmFlags, streams)
 	cmd := &cobra.Command{
-		Use:          "use",
-		Short:        "use change the current context to a cluster claim",
+		Use:          "set-cph",
+		Short:        "set-cph make the given clusterpoolhost active/current",
 		Example:      fmt.Sprintf(example, helpers.GetExampleHeader()),
 		SilenceUsage: true,
 		PreRunE: func(c *cobra.Command, args []string) error {